k-animator-project/utils/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
requests==2.31.0
azure-storage-blob==12.19.1
python-multipart==0.0.6
Pillow==10.3.0
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
azure-keyvault-secrets==4.10.0
azure-identity==1.25.0